refine make web api optional
