+++ mars/deploy/oscar/session.py
@@ -21,10 +21,10 @@ from weakref import WeakKeyDictionary
 from ...core import Tileable, enter_mode
 from ...core.session import AbstractSession, register_session_cls, \
     ExecutionInfo as AbstractExectionInfo, gen_submit_tileable_graph
-from ...services.meta import MetaAPI, OscarMetaAPI, WebMetaAPI
-from ...services.session import SessionAPI, OscarSessionAPI, WebSessionAPI
-from ...services.storage import OscarStorageAPI, WebStorageAPI
-from ...services.task import TaskAPI, OscarTaskAPI, WebTaskAPI, TaskResult
+from ...services.meta import MetaAPI, OscarMetaAPI
+from ...services.session import SessionAPI, OscarSessionAPI
+from ...services.storage import OscarStorageAPI
+from ...services.task import TaskAPI, OscarTaskAPI, TaskResult
 from ...utils import implements, merge_chunks
 from .typing import ClientType
 
@@ -80,14 +80,20 @@ class Session(AbstractSession):
             return (await new_cluster(address, **kwargs)).session
 
         use_web_api = urlparse(address).scheme == 'http'
-        session_api = await WebSessionAPI.create(address) if use_web_api else await OscarSessionAPI.create(address)
+        if use_web_api:
+            from ...services.session.web import WebSessionAPI
+            session_api = await WebSessionAPI.create(address)
+        else:
+            session_api = await OscarSessionAPI.create(address)
         # create new session
         session_address = await session_api.create_session(session_id)
         if use_web_api:
+            from ...services.meta.web import WebMetaAPI
             meta_api = await WebMetaAPI.create(address, session_id, session_address)
         else:
             meta_api = await OscarMetaAPI.create(session_id, session_address)
         if use_web_api:
+            from ...services.task.web import WebTaskAPI
             task_api = await WebTaskAPI.create(address, session_id, session_address)
         else:
             task_api = await OscarTaskAPI.create(session_id, session_address)
@@ -192,6 +198,7 @@ class Session(AbstractSession):
                 band = (await self._meta_api.get_chunk_meta(
                     chunk.key, fields=['bands']))['bands'][0]
                 if urlparse(self.address).scheme == 'http':
+                    from mars.services.storage.web import WebStorageAPI
                     storage_api = await WebStorageAPI.create(self.address, self._session_id, band[0])
                 else:
                     storage_api = await OscarStorageAPI.create(self._session_id, band[0])
