+++ mars/services/__init__.py
@@ -13,4 +13,3 @@
 # limitations under the License.
 
 from .core import NodeRole, start_services, stop_services
-from .web.core import get_supervisor_address
