+++ mars/services/meta/__init__.py
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .api import MetaAPI, MockMetaAPI, OscarMetaAPI, WebMetaAPI
+from .api import MetaAPI, MockMetaAPI, OscarMetaAPI
 from .core import TensorMeta, DataFrameMeta, SeriesMeta, IndexMeta, ObjectMeta, \
     TensorChunkMeta, DataFrameChunkMeta, SeriesChunkMeta, IndexChunkMeta, ObjectChunkMeta
