+++ mars/services/meta/web.py
@@ -12,8 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import List
+
 from mars.services.web.core import ServiceProxyHandlerBase, get_service_proxy_endpoint
-from .api import OscarMetaAPI
+from mars.services.web.core import ServiceWebAPIBase
+from ...utils import extensible
+from .api import MetaAPI, OscarMetaAPI
 
 
 class MetaAPIProxyHandler(ServiceProxyHandlerBase):
@@ -23,3 +27,17 @@ class MetaAPIProxyHandler(ServiceProxyHandlerBase):
 web_handlers = {
     get_service_proxy_endpoint('meta'): MetaAPIProxyHandler,
 }
+
+
+class WebMetaAPI(ServiceWebAPIBase, MetaAPI):
+    _service_name = 'meta'
+
+    @classmethod
+    async def create(cls, web_address: str, session_id: str, address: str):
+        return WebMetaAPI(web_address, 'create', session_id, address)
+
+    @extensible
+    async def get_chunk_meta(self,
+                             object_id: str,
+                             fields: List[str] = None):
+        return await self._call_method({}, 'get_chunk_meta', object_id, fields)
