+++ mars/services/session/__init__.py
@@ -12,5 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .api import SessionAPI, OscarSessionAPI, MockSessionAPI, WebSessionAPI
+from .api import SessionAPI, OscarSessionAPI, MockSessionAPI
 from .supervisor import SessionManagerActor, SessionActor
