+++ mars/services/session/web.py
@@ -12,8 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import Union
+
 from mars.services.web.core import ServiceProxyHandlerBase, get_service_proxy_endpoint
-from .api import OscarSessionAPI
+from mars.services.web.core import ServiceWebAPIBase, get_supervisor_address
+from .api import SessionAPI, OscarSessionAPI
 
 
 class SessionAPIProxyHandler(ServiceProxyHandlerBase):
@@ -23,3 +26,21 @@ class SessionAPIProxyHandler(ServiceProxyHandlerBase):
 web_handlers = {
     get_service_proxy_endpoint('session'): SessionAPIProxyHandler
 }
+
+
+class WebSessionAPI(ServiceWebAPIBase, SessionAPI):
+    _service_name = 'session'
+
+    @classmethod
+    async def create(cls, address: str, **kwargs):
+        supervisor_address = await get_supervisor_address(address)
+        return WebSessionAPI(address, 'create', supervisor_address, **kwargs)
+
+    async def create_session(self, session_id: str) -> str:
+        return await self._call_method({}, 'create_session', session_id)
+
+    async def delete_session(self, session_id: str):
+        return await self._call_method({}, 'delete_session', session_id)
+
+    async def get_last_idle_time(self, session_id: Union[str, None] = None) -> Union[float, None]:
+        return await self._call_method({}, 'get_last_idle_time', session_id)
