+++ mars/services/storage/web.py
@@ -12,8 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import Any, List
+
 from mars.services.web.core import ServiceProxyHandlerBase, get_service_proxy_endpoint
-from .api import OscarStorageAPI
+from mars.services.web.core import ServiceWebAPIBase, _transfer_request_timeout
+from ...storage.base import StorageLevel
+from .api import StorageAPI, OscarStorageAPI
+from .core import DataInfo
 
 
 class StorageAPIProxyHandler(ServiceProxyHandlerBase):
@@ -23,3 +28,20 @@ class StorageAPIProxyHandler(ServiceProxyHandlerBase):
 web_handlers = {
     get_service_proxy_endpoint('storage'): StorageAPIProxyHandler,
 }
+
+
+class WebStorageAPI(ServiceWebAPIBase, StorageAPI):
+    _service_name = 'storage'
+
+    @classmethod
+    async def create(cls, web_address: str, session_id: str, address: str):
+        return WebStorageAPI(web_address, 'create', session_id, address)
+
+    async def get(self, data_key: str, conditions: List = None) -> Any:
+        return await self._call_method(dict(request_timeout=_transfer_request_timeout),
+                                       'get', data_key, conditions)
+
+    async def put(self, data_key: str, obj: object,
+                  level: StorageLevel = StorageLevel.MEMORY) -> DataInfo:
+        return await self._call_method(dict(request_timeout=_transfer_request_timeout),
+                                       'put', data_key, obj, level)
